project: Code cleanup and reapply formatting
Changes applied:

* Moved utility files to /util/.
* Removed unused #includes.
* Removed unused ::ffmpeg::tools function.
* Removed unused variables.
* Fixed missing parentheses in the version macro.
* Fixed missing override on virtual function overrides and removed unnecessary virtual keyword from them.
* Disabled additional warning for ATL headers on MSVC only.
* Replaced direct printf parameters with their macro equivalent.
* Replaced C-style casts with C++-style casts.
* Applied clang-format again after an earlier change to the CMake file broke the integration for it.

+++ source/ffmpeg/hwapi/d3d11.cpp
@@ -25,7 +25,6 @@
 #include <sstream>
 #include <vector>
 #include "obs/gs/gs-helper.hpp"
-#include "utility.hpp"
 
 extern "C" {
 #pragma warning(push)
@@ -88,8 +87,9 @@ std::list<device> d3d11::enumerate_adapters()
 		dxgi_adapter->GetDesc1(&desc);
 
 		std::vector<char> buf(1024);
-		std::size_t       len = snprintf(buf.data(), buf.size(), "%ls (VEN_%04x/DEV_%04x/SUB_%04x/REV_%04x)",
-                                   desc.Description, desc.VendorId, desc.DeviceId, desc.SubSysId, desc.Revision);
+		std::size_t       len =
+			static_cast<size_t>(snprintf(buf.data(), buf.size(), "%ls (VEN_%04x/DEV_%04x/SUB_%04x/REV_%04x)",
+										 desc.Description, desc.VendorId, desc.DeviceId, desc.SubSysId, desc.Revision));
 
 		device dev;
 		dev.name      = std::string(buf.data(), buf.data() + len);
